feat: use enums for operators
Use enums and their variants as placeholders for the various operators
that can be used within the language. This allows for more controlled
handling in the parser that means the AST cannot have invalid operators.
This also fixes the implementation of the `=>` operator which previously
would be parsed as the `|` operator.